Refactor: `WalletInteraction`->`MetaMaskWalletInteraction`
