+++ src/modules/users/components/GasStation/MessageCardDetails/MessageCardControls.tsx
@@ -2,13 +2,11 @@ import { FormikProps } from 'formik';
 import React from 'react';
 import * as yup from 'yup';
 
-import { MessageType, WalletKind } from '~immutable/index';
+import { MessageType } from '~immutable/index';
 import Button from '~core/Button';
 import { ActionForm } from '~core/Fields';
-import WalletInteraction from '../WalletInteraction';
+import MetaMaskWalletInteraction from '../MetaMaskWalletInteraction';
 import { ActionTypes } from '~redux/index';
-import { useSelector } from '~utils/hooks';
-import { walletKindSelector } from '../../../selectors';
 import styles from './MessageCardControls.css';
 
 interface Props {
@@ -25,33 +23,28 @@ const validationSchema = yup.object().shape({
   id: yup.string(),
 });
 
-const MessageCardControls = ({ message: { id } }: Props) => {
-  const walletKind = useSelector(walletKindSelector);
-  return (
-    <div className={styles.main}>
-      <ActionForm
-        submit={ActionTypes.MESSAGE_SIGN}
-        success={ActionTypes.MESSAGE_SIGNED}
-        error={ActionTypes.MESSAGE_ERROR}
-        validationSchema={validationSchema}
-        initialValues={{ id }}
-      >
-        {({ isSubmitting }: FormikProps<FormValues>) => (
-          <Button
-            text={{ id: 'button.confirm' }}
-            type="submit"
-            loading={isSubmitting}
-          />
-        )}
-      </ActionForm>
-      {walletKind !== WalletKind.Software && (
-        <div className={styles.alert}>
-          <WalletInteraction walletKind={walletKind} />
-        </div>
+const MessageCardControls = ({ message: { id } }: Props) => (
+  <div className={styles.main}>
+    <ActionForm
+      submit={ActionTypes.MESSAGE_SIGN}
+      success={ActionTypes.MESSAGE_SIGNED}
+      error={ActionTypes.MESSAGE_ERROR}
+      validationSchema={validationSchema}
+      initialValues={{ id }}
+    >
+      {({ isSubmitting }: FormikProps<FormValues>) => (
+        <Button
+          text={{ id: 'button.confirm' }}
+          type="submit"
+          loading={isSubmitting}
+        />
       )}
+    </ActionForm>
+    <div className={styles.alert}>
+      <MetaMaskWalletInteraction />
     </div>
-  );
-};
+  </div>
+);
 
 MessageCardControls.displayName = displayName;
