+++ src/modules/users/components/GasStation/MetaMaskWalletInteraction/MetaMaskWalletInteraction.ts
@@ -1,0 +1,3 @@
+export const main: string;
+export const content: string;
+export const text: string;
