+++ src/modules/users/components/GasStation/MetaMaskWalletInteraction/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MetaMaskWalletInteraction';
