--- src/modules/users/components/GasStation/WalletInteraction/index.ts
@@ -1,1 +1,0 @@
-export { default } from './WalletInteraction';
